refactor : Google data class 네이밍 및 디렉토리 변경 #PAR-213
