--- core/network/src/main/java/online/partyrun/partyrunapplication/core/network/model/response/SignInGoogleResponse.kt
@@ -1,12 +1,0 @@
-package online.partyrun.partyrunapplication.core.network.model.response
-
-data class SignInGoogleResponse(
-    val userData: UserGoogleData?,
-    val errorMessage: String?
-)
-
-data class UserGoogleData(
-    val userId: String,
-    val username: String?,
-    val profilePictureUrl: String?
-)
