+++ feature/sign_in/src/main/java/online/partyrun/partyrunapplication/feature/sign_in/SignInViewModel.kt
@@ -9,10 +9,10 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import online.partyrun.partyrunapplication.core.model.signin.GoogleIdToken
-import online.partyrun.partyrunapplication.core.domain.GetSignInTokenUseCase
+import online.partyrun.partyrunapplication.core.domain.auth.GetSignInTokenUseCase
 import online.partyrun.partyrunapplication.core.common.network.ApiResponse
-import online.partyrun.partyrunapplication.core.domain.SaveTokensUseCase
-import online.partyrun.partyrunapplication.core.network.model.response.SignInGoogleResponse
+import online.partyrun.partyrunapplication.core.domain.auth.SaveTokensUseCase
+import online.partyrun.partyrunapplication.core.network.model.response.GoogleUserInfoResponse
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -25,7 +25,7 @@ class SignInViewModel @Inject constructor(
     private val _signInGoogleState = MutableStateFlow(SignInGoogleState())
     val signInGoogleState: StateFlow<SignInGoogleState> = _signInGoogleState.asStateFlow()
 
-    fun onSignInGoogleResult(result: SignInGoogleResponse) {
+    fun onSignInGoogleResult(result: GoogleUserInfoResponse) {
         _signInGoogleState.update {
             it.copy(
                 isSignInSuccessful = result.userData != null,
